requirements.txt
pyjwt
firebase-admin
pydantic